Add option to output data in the JSON format
