+++ matrixctl/addons/reports/addon.py
@@ -18,19 +18,18 @@
 
 from __future__ import annotations
 
+import json
 import logging
 
 from argparse import Namespace
-from shutil import get_terminal_size
-from textwrap import TextWrapper
 
 from matrixctl.errors import InternalResponseError
 from matrixctl.handlers.api import RequestBuilder
 from matrixctl.handlers.api import request
-from matrixctl.handlers.table import table
 from matrixctl.handlers.yaml import YAML
-from matrixctl.print_helpers import timestamp_to_dt
 from matrixctl.typehints import JsonDict
+
+from .to_table import to_table
 
 
 __author__: str = "Michael Sasser"
@@ -40,43 +39,12 @@ __email__: str = "Michael@MichaelSasser.org"
 logger = logging.getLogger(__name__)
 
 
-def addon(_: Namespace, yaml: YAML) -> int:
-    """Print a table of the reported events.
-
-    Examples
-    --------
-    .. code-block:: console
-
-       $ matrixctl reports
-       +-----------------+----------------------------------------------+
-       | ID              | 2                                            |
-       | Date            | 2021-05-08                                   |
-       | Time            | 21:04:55                                     |
-       | Score           | -100                                         |
-       | Canonical Alias | -                                            |
-       | Room Name       | SomeRoom                                     |
-       | Room ID         | !AbCdEfGhIjKlMnOpQr:domain.tld               |
-       | Event ID        | $Q_sksd348jaidj93jf9ojwef9h329ofijewhf932h9f |
-       | Defendant       | @mallory:matrix.org                          |
-       | Plaintiff       | @alice:myhomeverver.tld                      |
-       | Reason          | Likes JavaScript                             |
-       |-----------------+----------------------------------------------|
-       | ID              | 1                                            |
-       | Date            | 2020-08-15                                   |
-       | Time            | 09:09:57                                     |
-       | Score           | -100                                         |
-       | Canonical Alias | -                                            |
-       | Room Name       | -                                            |
-       | Room ID         | !AbCdEfGhIjKlMnOpQr:matrix.org               |
-       | Event ID        | $123456789012345678901:matrix.org            |
-       | Defendant       | @eve:matrix.org                              |
-       | Plaintiff       | @bob:myhomeserver.tld                        |
-       | Reason          | Hates The Office (US)                        |
-       +-----------------+----------------------------------------------+
+def addon(arg: Namespace, yaml: YAML) -> int:
+    """Print a table/json of the reported events.
 
     Parameters
     ----------
-    _ : argparse.Namespace
+    arg : argparse.Namespace
         The ``Namespace`` object of argparse's ``parse_args()``.
     yaml : matrixctl.handlers.yaml.YAML
         The configuration file handler.
@@ -115,59 +83,11 @@ def addon(_: Namespace, yaml: YAML) -> int:
         except KeyError:
             break
 
-    # pprint(events_raw)
-    events: list[tuple[str, str]] = []
-
-    logger.debug(f"Terminal width = {get_terminal_size().columns}")
-
-    wrapper_reason = TextWrapper(
-        width=get_terminal_size().columns - 20,  # is const.
-        drop_whitespace=True,
-        break_long_words=True,
-    )
-
-    for event in events_raw:
-        dt: str = timestamp_to_dt(event["received_ts"], "\n")
-        canonical_alias: str = (
-            event["canonical_alias"]
-            if event["canonical_alias"] is not None
-            else "-"
-        )
-
-        events.append(
-            (
-                "\n".join(
-                    (
-                        "ID",
-                        "Date",
-                        "Time",
-                        "Score",
-                        "Canonical Alias",
-                        "Room Name",
-                        "Room ID",
-                        "Event ID",
-                        "Defendant",
-                        "Plaintiff",
-                        "Reason",
-                    )
-                ),
-                (
-                    f"{event['id']}\n"
-                    f"{dt}\n"
-                    f"{event['score']}\n"
-                    f"{canonical_alias}\n"
-                    f"{event['name'] if event['name'] else '-'}\n"
-                    f"{event['room_id']}\n"
-                    f"{event['event_id']}\n"
-                    f"{event['sender']}\n"
-                    f"{event['user_id']}\n"
-                    f"{wrapper_reason.fill(text=event['reason'])}"
-                ),
-            )
-        )
-
-    for line in table(events):
-        print(line)
+    if arg.to_json:
+        print(json.dumps(events_raw, indent=4))
+    else:
+        for line in to_table(events_raw):
+            print(line)
 
     return 0
