+++ news/211.feature
@@ -1,0 +1,1 @@
+With the ``-j`` or ``--to-json`` argument, the output of some addons can be set to the JSON format.
